requirements.txt
google-adk
usearch
numpy
Pillow
